Separate interface and implementation into separate files.
